Remove /W3 warnings

+++ SpectralEngine/Common/Camera.cpp
@@ -90,7 +90,7 @@ float Camera::GetFovY()const
 float Camera::GetFovX()const
 {
 	float halfWidth = 0.5f*GetNearWindowWidth();
-	return 2.0f*atan(halfWidth / mNearZ);
+	return 2.0f * static_cast<float>(atan(halfWidth / mNearZ));
 }
 
 float Camera::GetNearWindowWidth()const
